Add files via upload
The Castle in the Air is designed to be a thorough test of the capabilities of both you and your printer, as well as a method to better understand the various print settings and their impact on part quality. In addition to testing a wide variety of part features, these also test assembly fit tolerances by interlocking and stacking together.

Score your print and compare your results at http://www.printbetter.parts !

Share your print and tag it with #PrintTheCastle!

ONLY TO BE ATTEMPTED BY THE BRAVE AT HEART; this is not intended to print perfectly the first time, as it will reveal flaws in most machines and slicers. To aid in your attempts, a scoring quiz has been devised to help navigate this quest. The scoring is tiered, with each level out of 10, one point per perfect feature. Your ranking will increase From Jester → Wizard depending on your parts abilities.

This piece will test the quality of a printed object with bridging, small features, flat surfaces, curved surfaces, stringing, and internal channels. The whole castle should take 1.5 → 3 hours to print and use ≈25g of filament.