Add files via upload
This is a test chain printed in one solid assembly (no support or bridging) that creates a solid, flexible chain. There are 5 links in this chain, each able to move independently of the others. The part prints in one solid print, which comes of the bed as a semi-stiff chain. With a few bends, the chain loosens and becomes a flexible system.