Add files via upload
This is a test object to calibrate bridges on your printer, and is designed to print as quickly and economically as possible, by minimizing time and material spent on areas not necessary for proper calibration.
Bridging is your printer's ability to print a layer between gaps in lower layers without support, essentially printing over thin air, creating a "bridge". It is very important that your printer can properly bridge, because many prints require some bridging.
Head on over to the 'instructions' tab to get started.
The first two files (pictured) have an overall length is 40mm with a bridge 30mm long.
The first file is about 13mm (1/2") tall, and the second file is about 6mm (1/4") tall to save even more time.
I also added two more bridges, one with a 75mm bridge, and the other with a whopping 100mm bridge!