Add files via upload
Cali Cat is designed to be the purrfect initial calibration and benchmarking print. It is a simple, standard test to ensure you have the basics down before proceeding to advanced prints. Cali Cat is the alternative to boring calibration cubes. It's a test print you can enjoy keeping. It consists of linear dimensioning tests of a calibration cube and more:


Linear dimensioning: Body 20x20mm, Height: 35mm, Tail 5x5mm
Inclines and overhangs: 45º
Prints in under 1 hour.

Calibrate your printer and grow an army of cats in the process!